requirements.txt
pdfplumber
PyMuPDF
pdf2image
pytesseract
pypdf
Pillow
